Bone data compiled and loaded successfully, need to print the data to verify

+++ src/some_data.rs
@@ -199,5 +199,5 @@ pub const TREES:[&str; 6] = [
     "resources/models/obj/tree_oak_dark.obj",
 ];
 
-pub const MAX_BONE_INFLUENCE: u32 = 4;
+pub const MAX_BONE_INFLUENCE: usize = 4;
 pub const MAX_BONES: u32 = 200;
